Finished RuneLite HD

+++ runelite-client/src/main/java/net/runelite/client/plugins/hd/materials/GroundMaterial.java
@@ -105,7 +105,7 @@ public enum GroundMaterial
 		// Generate a seed from the tile coordinates for
 		// consistent 'random' results between scene loads.
 		// This seed creates a patchy, varied terrain
-		long seed = (long) (plane + 1) * 10 * (worldX % 100) * 20 * (worldY % 100) * 30;
+		long seed = (plane + 1) * 10 * (worldX % 100) * 20 * (worldY % 100) * 30;
 		randomTex.setSeed(seed);
 		int randomInt = randomTex.nextInt(this.materials.length);
 		return this.materials[randomInt];
